Better error handling and resource management

+++ src/main/java/ca/corbett/extras/demo/panels/ActionPanelDemoPanel.java
@@ -31,6 +31,7 @@ import ca.corbett.forms.fields.PanelField;
 import ca.corbett.forms.fields.ShortTextField;
 import ca.corbett.forms.validators.ValidationResult;
 
+import javax.sound.sampled.AudioInputStream;
 import javax.sound.sampled.AudioSystem;
 import javax.sound.sampled.Clip;
 import javax.swing.AbstractAction;
@@ -161,23 +162,49 @@ public class ActionPanelDemoPanel extends PanelBuilder implements ExpandListener
         ((CardLayout)cardPanel.getLayout()).show(cardPanel, TAB_INTRO);
 
         // Just for fun, load some sound effects:
+        AudioInputStream expandInputStream = null;
+        AudioInputStream collapseInputStream = null;
         try {
             // AudioSystem will puke if we try to load directly from a resource stream,
             // so we have to extract to a temp file first. Fun!
             File tmpExpand = File.createTempFile("expand", ".wav");
             tmpExpand.deleteOnExit();
-            ResourceLoader.extractResourceToFile(EXPAND_CLIP, tmpExpand);
+            if (!ResourceLoader.extractResourceToFile(EXPAND_CLIP, tmpExpand)) {
+                throw new Exception("Failed to extract resource " + EXPAND_CLIP + " to temp file");
+            }
             File tmpCollapse = File.createTempFile("collapse", ".wav");
             tmpCollapse.deleteOnExit();
-            ResourceLoader.extractResourceToFile(COLLAPSE_CLIP, tmpCollapse);
+            if (!ResourceLoader.extractResourceToFile(COLLAPSE_CLIP, tmpCollapse)) {
+                throw new Exception("Failed to extract resource " + COLLAPSE_CLIP + " to temp file");
+            }
             expandClip = AudioSystem.getClip();
-            expandClip.open(AudioSystem.getAudioInputStream(tmpExpand));
+            expandInputStream = AudioSystem.getAudioInputStream(tmpExpand);
+            expandClip.open(expandInputStream);
             collapseClip = AudioSystem.getClip();
-            collapseClip.open(AudioSystem.getAudioInputStream(tmpCollapse));
+            collapseInputStream = AudioSystem.getAudioInputStream(tmpCollapse);
+            collapseClip.open(collapseInputStream);
         }
         catch (Exception e) {
             // No fun for you!
             log.log(Level.SEVERE, "Could not load sound effects :(", e);
+        }
+        finally {
+            if (expandInputStream != null) {
+                try {
+                    expandInputStream.close();
+                }
+                catch (Exception e) {
+                    log.log(Level.WARNING, "Failed to close audio input stream for expand clip");
+                }
+            }
+            if (collapseInputStream != null) {
+                try {
+                    collapseInputStream.close();
+                }
+                catch (Exception e) {
+                    log.log(Level.WARNING, "Failed to close audio input stream for collapse clip");
+                }
+            }
         }
 
         // Register for L&F updates so we can keep our demo panels updated if the user changes L&F:
